Refactor carma_wm to better support unit testing
