+++ carma_wm/test/IndexedDistanceMapTest.cpp
@@ -16,7 +16,7 @@
 
 #include <gmock/gmock.h>
 #include <iostream>
-#include <../src/CARMAWorldModel.h>
+#include <carma_wm/CARMAWorldModel.h>
 #include <lanelet2_core/geometry/LineString.h>
 #include "TestHelpers.h"
