+++ carma_wm/test/TestHelpers.h
@@ -17,7 +17,7 @@
 
 #include <gmock/gmock.h>
 #include <iostream>
-#include <../src/CARMAWorldModel.h>
+#include <carma_wm/CARMAWorldModel.h>
 #include <lanelet2_core/geometry/LineString.h>
 #include <lanelet2_traffic_rules/TrafficRulesFactory.h>
 #include <lanelet2_core/Attribute.h>
