+++ carma_wm/test/WMListenerWorkerTest.cpp
@@ -18,7 +18,7 @@
 #include <iostream>
 #include <lanelet2_extension/utility/message_conversion.h>
 #include <../src/WMListenerWorker.h>
-#include <../src/CARMAWorldModel.h>
+#include <carma_wm/CARMAWorldModel.h>
 #include <lanelet2_core/geometry/LineString.h>
 #include <lanelet2_traffic_rules/TrafficRulesFactory.h>
 #include <lanelet2_core/Attribute.h>
